move Coords2 tests to dedicated module
